remove dll in favor of auto-build
