+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.1] - 2025-04-11
+
+- rebuilt mod for 7dtd 1.4 (b8)
+- update pipeline to build multiple versions
+
 ## [5.0.0] - 2025-02-15
 
 - update references for 7dtd-1.3-b9
