+++ ModInfo.xml
@@ -4,7 +4,7 @@
     <DisplayName value="Quantum Elevators" />
     <!-- REMINDER: this mod's DLL version will be logged on GameAwake and should reflect the same version listed here! -->
     <!--  If troubleshooting an issue, search "Quantum Elevators DLL" in logs to confirm these two values match. -->
-    <Version value="5.0.0" />
+    <Version value="5.0.1" />
     <Description value="Add infinite distance, vertical-warp elevator panels." />
     <Author value="Jonathan Robertson (Kanaverum)" />
     <Website value="https://github.com/jonathan-robertson/quantum-elevators" />
